Add files via upload
A to do list is a list of tasks that contains everything you need to complete. It ensures you have everything written down in one place so you don't forget anything important. These lists don't have to be that complicated. You can use a Post-It note or even a piece of scrap paper to write down your tasks.